Expiration check w/ server time check
