+++ package.json
@@ -22,7 +22,6 @@
     "react-contextmenu": "^2.9.2",
     "react-dom": "^16.2.0",
     "react-dropzone": "^4.2.11",
-    "react-facebook-login": "^4.0.1",
     "react-google-login": "^3.2.1",
     "react-material-ui-form-validator": "^2.0.0-beta.4",
     "react-player": "^1.6.0",
