--- src/components/AudioPlayer/styles/AudioPlayer.css
@@ -1,5 +1,0 @@
-.player-container {
-  position: fixed;
-  bottom: 10px;
-  right: 14px;
-}
