+++ src/components/Explorer/NodeViewer/AudioPlayer/styles/AudioPlayer.less
@@ -1,0 +1,5 @@
+.player-container {
+  position:fixed;
+  bottom: 10px;
+  right : 14px;
+}
