+++ src/components/Explorer/NodeViewer/UploadView/styles/UploadView.css
@@ -1,0 +1,43 @@
+.hidden {
+  display: none;
+}
+.dl-header {
+  background-color: #353535;
+  color: white;
+  padding: 10px;
+}
+.dl-header .header-text {
+  display: flex;
+  align-items: center;
+}
+.dl-header .header-icons .button {
+  color: #9e9e9e;
+}
+.dl-header .header-icons .button:hover {
+  color: white;
+}
+.dl-item {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  padding: 10px;
+}
+.loader {
+  margin-right: 25px;
+  border: 3px solid #f3f3f3;
+  /* Light grey */
+  border-top: 3px solid #3498db;
+  /* Blue */
+  border-radius: 50%;
+  width: 20px;
+  height: 20px;
+  animation: spin 2s linear infinite;
+}
+@keyframes spin {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
