+++ src/components/Explorer/NodeViewer/UploadView/styles/UploadView.less
@@ -1,0 +1,44 @@
+.hidden {
+  display: none;
+}
+
+.dl-header {
+  background-color: #353535;
+  color : white;
+  padding: 10px;
+
+  .header-text {
+    display: flex;
+    align-items: center;
+  }
+  .header-icons {
+    .button {
+      color: #9e9e9e;
+      &:hover {
+        color: white;
+      }
+    }
+  }
+}
+
+.dl-item {
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  padding: 10px;
+}
+
+.loader {
+  margin-right: 25px;
+  border: 3px solid #f3f3f3; /* Light grey */
+  border-top: 3px solid #3498db; /* Blue */
+  border-radius: 50%;
+  width: 20px;
+  height: 20px;
+  animation: spin 2s linear infinite;
+}
+
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
