--- src/components/Login/Signup/styles/Signup.less
@@ -1,83 +1,0 @@
-@import '../../../libs/lib.less';
-
-.card {
-  font-family: 'Montserrat', sans-serif;
-  width:400px;
-  height: 510px;
-  margin: 100px auto 8px;
-  padding-bottom: 20px;
-  background-color: #fff;
-  border-color: #ddd #ddd #d8d8d8;
-  border-radius: 3px;
-  border-style: solid;
-  border-width: 1px;
-  box-shadow: 0 2px 0 rgba(175, 175, 175, 0.12) !important;
-  .header {
-    .logoContainer {
-      margin: 30px auto;
-      padding-bottom: 30px;
-      position: relative;
-      border-bottom: 1px solid #ddd;
-      .logoIcon {
-        display: block;
-        margin: 0 auto;
-        max-width: 200px;
-        max-height: 40px;
-      }
-    }
-
-  }
-  .cardContent{
-    padding : 20px;
-    .form-head{
-      font-size: 14px;
-      line-height: 1.5;
-      margin-top: 10px;
-      margin-bottom: 0px;
-      text-align: center;
-      text-transform: none;
-      font-weight: bold;
-      color: #5e5e5e;
-  
-    }
-    .form-container {
-      padding-left: 20px;
-      padding-right: 20px;
-      display: flex;
-      flex-wrap: wrap;
-
-      .button {
-        font-family: 'Montserrat', sans-serif;
-        font-size : 13px;
-        margin-top : 75px;
-        height: 50px;
-        text-transform: capitalize;
-        line-height: 36px;
-        border-radius: 3px;
-        border-style: solid;
-        border-width: 1px;
-        border-color: @p-dark;
-        &:hover {
-          background-color: @p-light;
-        }
-      }
-  
-    };
-  }
-
-};
-
-@media all and (max-width: 480px){
-  .card {
-    width : 300px;
-  }  
-}
-
-@media all and (max-height: 760px){
-  #okta-sign-in, .card {
-      margin-top: 0;
-  }
-  .tabsHeader{
-    top : 10px !important;
-  }
-}
