--- src/components/Login/styles/Login.less
@@ -1,56 +1,0 @@
-@import '../../../libs/lib.less';
-
-.forms-container {
-  max-height: 90vh;
-  overflow-y: scroll;
-  padding-bottom: 100px;
-  .tabsHeader {
-    position: relative !important;
-    overflow: auto;
-    border-radius: 3px;
-    border-style: solid;
-    border-width: 1px;
-    height: auto;
-    width: 405px;
-    min-width: 300px;
-    top: 105px;
-    margin: 0 auto 8px;
-  
-    -webkit-box-shadow: none !important;
-    -moz-box-shadow: none !important;
-    box-shadow: none !important;
-    
-    border-color: #ddd #ddd #d8d8d8;
-    button {
-      width : 50%
-    }
-  
-  }
-
-}
-
-#okta-sign-in.auth-container.main-container {
-  margin-top: 100px;
-  margin-right: auto;
-  margin-bottom: 8px;
-  margin-left: auto;
-}
-
-
-@media all and ( max-width : 480px) {
-  .forms-container {
-    .tabsHeader {
-      width:305px;
-    }
-    #okta-sign-in.auth-container.main-container {
-      width:300px;
-    }
-  }
-}
-
-
-@media all and ( max-height : 760px) {
-  #okta-sign-in.auth-container.main-container {
-    margin-top: 0px !important;
-  }
-}
